build: add trailing newline when patching Dockerfile
When check-container-build-with-prci.sh patches the Dockerfile, it
does not append a trailing newline to the new COPY line.  If there
is an empty line following the FROM directive, this is fine.  But
otherwise, it breaks the Dockerfile.

Append a trailing newline so that adjacent lines do not get joined
by accident.