build: avoid cache invalidation due to repo file
Build caching is aware of not only the content of files that are
added to the image via COPY, but also the mtime.  Therefore using
curl to fetch the repo file defeats build caching.

To avoid spurious cache invalidation, download the repo file to a
temp file.  Only replace the existing repo file if the content
changed, or if user demands is via $FORCE.